requirements.txt
aniso8601==9.0.1
asn1crypto==0.24.0
autochain==6.2.1
Babel==2.5.1
bcrypt==3.2.0
bidict==0.21.4
bitstruct==8.15.1
cassandra-driver==3.11.0
#certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
#charset-normalizer==3.1.0
click==6.7
#contourpy==1.0.7
crc8==0.1.0
#cryptography==38.0.4
#cycler==0.10.0
daemontool==4.1.0
decorator==4.4.2
dnspython==2.1.0
et-xmlfile==1.1.0
file-magic==0.4.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-RESTful==0.3.6
#fonttools==4.39.0
frugy==0.3.1
futures==3.1.1
gunicorn==19.7.1
idna==3.4
#imageio==2.9.0
itsdangerous==0.24
Jinja2==2.9.6
kiwisolver==1.4.4
MarkupSafe==2.1.2
#matplotlib==3.7.1
#networkx==2.5.1
numpy==1.23.5
olefile==0.46
openpyxl==3.1.2
#packaging==23.0
#pandas==1.2.4
paramiko==2.8.1
Pillow==9.4.0
pip==23.0.1
ply==3.11
#pycparser==2.18
PyNaCl==1.5.0
pyocr==0.8
#pyOpenSSL==22.1.0
#pyparsing==3.0.0b2
pyserial==3.5
#PySocks==1.6.7
python-dateutil==2.8.2
#pytz==2017.3
PyWavelets==1.4.1
PyYAML==6.0
#requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
#scikit-image==0.18.1
#scipy==1.10.1
setuptools==65.6.3
setuptools-scm==6.0.1
#six==1.11.0
tabulate==0.8.10
#tifffile==2020.9.3
toml==0.10.2
tqdm==4.65.0
urllib3==1.26.14
urwid==2.0.1
Werkzeug==0.12.2
wheel==0.38.4
requests_ntlm==1.2.0
pydot==1.4.2